Refactor iterator usage in benchmarks
Replaced manual index iteration with iterator methods for creating
`Vec<SmallNode>` and iterating over nodes, improving code readability
and conciseness. This change also slightly alters the iteration logic in
`bench_steady_state` for clarity.

Additionally, a minor change was made in `bench_pop_clustered` to ensure
the `size` variable is of type `u64` for consistent
`Throughput::Elements` usage.

+++ benches/released_set.rs
@@ -243,8 +243,8 @@ fn bench_pop(c: &mut Criterion) {
 fn bench_pop_clustered(c: &mut Criterion) {
     let mut group = c.benchmark_group("released_set/pop_clustered");
 
-    let size = 1000;
-    group.throughput(Throughput::Elements(size as u64));
+    let size = 1000u64;
+    group.throughput(Throughput::Elements(size));
 
     // Sequential keys (may cluster based on hash distribution)
     let sequential_keys: Vec<u64> = (0..size).collect();
